requirements.txt
pandas
streamlit
matplotlib
numpy
seaborn
Pillow
mpld3
hydralit_components